# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "minotari",
    "description": "",
    "contact": {
      "name": "The Tari Development Community"
    },
    "license": {
      "name": "BSD-3-Clause",
      "identifier": "BSD-3-Clause"
    },
    "version": "0.2.0"
  },
  "paths": {
...
    "/accounts/{name}/burn": {
      "post": {
        "tags": [
          "accounts"
        ],
        "summary": "Burns funds from an account and records a partial burn proof for L2 claiming.",
        "description": "Creates a burn transaction, broadcasts it, and stores the partial proof in the\ndatabase. The daemon's `BurnProofWorker` will automatically fetch the kernel\nmerkle proof once the transaction is confirmed and write the complete\n`CompleteClaimBurnProof` JSON file to the configured `burn_proofs_dir`.\n\n# Path Parameters\n\n- `name`: The account to burn funds from\n\n# Request Body\n\nSee [`BurnFundsRequest`] for the complete schema.\n\n# Response\n\nReturns a [`BurnFundsResponse`] with:\n- `tx_id`: the transaction ID\n- `output_hash`: hex-encoded output hash (use this to identify the proof file)\n\n# Errors\n\n- `400 Bad Request`: invalid `claim_public_key` or `sidechain_deployment_key` hex\n- `404 Not Found`: account does not exist\n- `500`: insufficient funds, broadcast failure, or other internal error\n\n# Example Request\n\n```bash\ncurl -X POST http://localhost:9000/accounts/default/burn \\\n  -H \"Authorization: Bearer $MINOTARI_API_TOKEN\" \\\n  -H \"Content-Type: application/json\" \\\n  -d '{\"amount\": 1000000, \"claim_public_key\": \"a3f9...\"}'\n```",
        "operationId": "api_burn_funds",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "description": "Name of the account to burn from",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BurnFundsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Burn transaction broadcast successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BurnFundsResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Account not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Burn failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
    "/accounts/{name}/events": {
      "get": {
        "tags": [
          "accounts"
        ],
        "summary": "Retrieves wallet events for a specified account with pagination.",
        "description": "Returns a paginated list of events that have occurred for the account, including\noutput detection, confirmation, transaction broadcasts, and blockchain\nreorganizations. Events are ordered by creation time (most recent first).\n\n# Path Parameters\n\n- `name`: The unique account name to query\n\n# Query Parameters\n\n- `limit`: Maximum number of events to return (default: 50, max: 1000)\n- `offset`: Number of events to skip for pagination (default: 0)\n\n# Response\n\nReturns a list of [`DbWalletEvent`] objects, each containing:\n- Event ID and type\n- Human-readable description\n- JSON data with event-specific details\n- Creation timestamp\n\n# Errors\n\n- [`ApiError::AccountNotFound`]: The specified account does not exist\n- [`ApiError::DbError`]: Database connection or query failure\n\n# Example Request\n\n```bash\n# Get first 50 events (default)\ncurl -X GET http://localhost:8080/accounts/default/events \\\n  -H \"Authorization: Bearer $MINOTARI_API_TOKEN\"\n\n# Get 100 events starting from offset 50\ncurl -X GET \"http://localhost:8080/accounts/default/events?limit=100&offset=50\" \\\n  -H \"Authorization: Bearer $MINOTARI_API_TOKEN\"\n```\n\n# Example Response\n\n```json\n[\n  {\n    \"id\": 42,\n    \"account_id\": 1,\n    \"event_type\": \"OutputDetected\",\n    \"description\": \"Detected output at height 12345\",\n    \"data_json\": \"{\\\"hash\\\":\\\"abc...\\\",\\\"block_height\\\":12345}\",\n    \"created_at\": \"2024-01-15T10:30:00\"\n  }\n]\n```",
        "operationId": "api_get_events",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "description": "Name of the account to retrieve events for",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Maximum number of events to return (default: 50, max: 1000)",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "description": "Number of events to skip for pagination (default: 0)",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Account events retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DbWalletEvent"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Account not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiError": {
        "oneOf": [
          {
            "type": "object",
            "description": "A general internal server error with a descriptive message.\n\nUsed for unexpected errors that don't fit other categories.\nReturns HTTP 500 Internal Server Error.",
            "required": [
              "InternalServerError"
            ],
            "properties": {
              "InternalServerError": {
                "type": "string",
                "description": "A general internal server error with a descriptive message.\n\nUsed for unexpected errors that don't fit other categories.\nReturns HTTP 500 Internal Server Error."
              }
            }
          },
          {
            "type": "object",
            "description": "A database operation failed.\n\nThis includes connection failures, query errors, and constraint\nviolations. Returns HTTP 500 Internal Server Error.\n\n# Example Causes\n\n- Database connection pool exhausted\n- SQL query syntax error\n- Foreign key constraint violation",
            "required": [
              "DbError"
            ],
            "properties": {
              "DbError": {
                "type": "string",
                "description": "A database operation failed.\n\nThis includes connection failures, query errors, and constraint\nviolations. Returns HTTP 500 Internal Server Error.\n\n# Example Causes\n\n- Database connection pool exhausted\n- SQL query syntax error\n- Foreign key constraint violation"
              }
            }
          },
          {
            "type": "object",
            "description": "The requested account was not found.\n\nThe contained string is the account name that was not found.\nReturns HTTP 404 Not Found.\n\n# Example\n\n```json\n{\n  \"error\": \"Account 'nonexistent' not found\"\n}\n```",
            "required": [
              "AccountNotFound"
            ],
            "properties": {
              "AccountNotFound": {
                "type": "string",
                "description": "The requested account was not found.\n\nThe contained string is the account name that was not found.\nReturns HTTP 404 Not Found.\n\n# Example\n\n```json\n{\n  \"error\": \"Account 'nonexistent' not found\"\n}\n```"
              }
            }
          },
          {
            "type": "object",
            "description": "A requested resource was not found.\n\nUsed for generic not found errors beyond just accounts.\nReturns HTTP 404 Not Found.\n\n# Example\n\n```json\n{\n  \"error\": \"No blocks have been scanned yet\"\n}\n```",
            "required": [
              "NotFound"
            ],
            "properties": {
              "NotFound": {
                "type": "string",
                "description": "A requested resource was not found.\n\nUsed for generic not found errors beyond just accounts.\nReturns HTTP 404 Not Found.\n\n# Example\n\n```json\n{\n  \"error\": \"No blocks have been scanned yet\"\n}\n```"
              }
            }
          },
          {
            "type": "object",
            "description": "A bad request error for invalid input.\n\nUsed when the client sends invalid data in the request.\nReturns HTTP 400 Bad Request.\n\n# Example\n\n```json\n{\n  \"error\": \"Invalid hex in payment_id_hex: Invalid character\"\n}\n```",
            "required": [
              "BadRequest"
            ],
            "properties": {
              "BadRequest": {
                "type": "string",
                "description": "A bad request error for invalid input.\n\nUsed when the client sends invalid data in the request.\nReturns HTTP 400 Bad Request.\n\n# Example\n\n```json\n{\n  \"error\": \"Invalid hex in payment_id_hex: Invalid character\"\n}\n```"
              }
            }
          },
          {
            "type": "object",
            "description": "Failed to lock funds for a transaction.\n\nThis typically occurs when there are insufficient available funds\nor when UTXO selection fails. Returns HTTP 500 Internal Server Error.\n\n# Common Causes\n\n- Insufficient balance in the account\n- All UTXOs are already locked by other pending transactions\n- UTXO selection algorithm could not find suitable inputs",
            "required": [
              "FailedToLockFunds"
            ],
            "properties": {
              "FailedToLockFunds": {
                "type": "string",
                "description": "Failed to lock funds for a transaction.\n\nThis typically occurs when there are insufficient available funds\nor when UTXO selection fails. Returns HTTP 500 Internal Server Error.\n\n# Common Causes\n\n- Insufficient balance in the account\n- All UTXOs are already locked by other pending transactions\n- UTXO selection algorithm could not find suitable inputs"
              }
            }
          },
          {
            "type": "object",
            "description": "Failed to create an unsigned transaction.\n\nThis occurs during transaction construction after funds have been\nlocked. Returns HTTP 500 Internal Server Error.\n\n# Common Causes\n\n- Invalid recipient address format\n- Transaction size exceeds limits\n- Cryptographic operation failure",
            "required": [
              "FailedCreateUnsignedTx"
            ],
            "properties": {
              "FailedCreateUnsignedTx": {
                "type": "string",
                "description": "Failed to create an unsigned transaction.\n\nThis occurs during transaction construction after funds have been\nlocked. Returns HTTP 500 Internal Server Error.\n\n# Common Causes\n\n- Invalid recipient address format\n- Transaction size exceeds limits\n- Cryptographic operation failure"
              }
            }
          },
          {
            "type": "object",
            "description": "Failed to build or broadcast a burn transaction.\n\nThis occurs during burn transaction construction or network submission.\nReturns HTTP 500 Internal Server Error.\n\n# Common Causes\n\n- Insufficient balance\n- Invalid claim public key\n- Base node rejected the transaction",
            "required": [
              "FailedToBurnFunds"
            ],
            "properties": {
              "FailedToBurnFunds": {
                "type": "string",
                "description": "Failed to build or broadcast a burn transaction.\n\nThis occurs during burn transaction construction or network submission.\nReturns HTTP 500 Internal Server Error.\n\n# Common Causes\n\n- Insufficient balance\n- Invalid claim public key\n- Base node rejected the transaction"
              }
            }
          }
        ],
        "description": "Represents all possible errors returned by the REST API.\n\nEach variant corresponds to a specific error condition that can occur\nduring API request processing. The error type automatically converts\nto an appropriate HTTP response with a JSON error body.\n\n# Error Handling Pattern\n\nAPI handlers typically use the `?` operator with this error type:\n\n```rust,ignore\npub async fn handler() -> Result<Json<Data>, ApiError> {\n    let data = fetch_data().await?; // Errors automatically convert to ApiError\n    Ok(Json(data))\n}\n```\n\n# Serialization\n\nWhen serialized to JSON for API responses, errors produce:\n\n```json\n{\n  \"error\": \"Error message here\"\n}\n```"
      },
...
      "BurnFundsRequest": {
        "type": "object",
        "description": "Request body for burning funds and generating an L2 claim proof.\n\n# JSON Example\n\n```json\n{\n  \"amount\": 1000000,\n  \"claim_public_key\": \"a3f9...\",\n  \"fee_per_gram\": 5,\n  \"payment_id\": \"optional memo\",\n  \"seconds_to_lock\": 86400\n}\n```",
        "required": [
          "amount"
        ],
        "properties": {
          "amount": {
            "type": "integer",
            "format": "int64",
            "description": "Amount to burn in MicroMinotari.",
            "minimum": 0
          },
          "claim_public_key": {
            "type": [
              "string",
              "null"
            ],
            "description": "L2 claim public key in hex. Required to generate a burn proof for L2 claiming."
          },
          "fee_per_gram": {
            "type": "integer",
            "format": "int64",
            "description": "Fee per gram in MicroMinotari (default: 5).",
            "minimum": 0
          },
          "idempotency_key": {
            "type": [
              "string",
              "null"
            ],
            "description": "Optional idempotency key to prevent duplicate burn requests."
          },
          "payment_id": {
            "type": [
              "string",
              "null"
            ],
            "description": "Optional payment memo attached to the transaction."
          },
          "seconds_to_lock": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "Seconds to lock input UTXOs (default: 86400 = 24 h).",
            "minimum": 0
          },
          "sidechain_deployment_key": {
            "type": [
              "string",
              "null"
            ],
            "description": "Sidechain deployment key in hex, for L2 template burns."
          }
        }
      },
      "BurnFundsResponse": {
        "type": "object",
        "description": "Response returned after a successful burn transaction.",
        "required": [
          "tx_id",
          "output_hash"
        ],
        "properties": {
          "output_hash": {
            "type": "string",
            "description": "Hex-encoded output hash of the burn output.\n\nThe daemon will use this to track when the proof is ready."
          },
          "tx_id": {
            "type": "string",
            "description": "Transaction ID assigned to the burn transaction."
          }
        }
      },
...
      "DbWalletEvent": {
        "type": "object",
        "description": "A database row representation of a wallet event.\n\nThis struct is used for deserializing wallet events from the database\nand serializing them for API responses.",
        "required": [
          "id",
          "account_id",
          "event_type",
          "description",
          "created_at"
        ],
        "properties": {
          "account_id": {
            "type": "integer",
            "format": "int64",
            "description": "Account this event belongs to"
          },
          "created_at": {
            "type": "string",
            "description": "Timestamp when the event was created"
          },
          "data_json": {
            "type": [
              "string",
              "null"
            ],
            "description": "JSON data containing event-specific details"
          },
          "description": {
            "type": "string",
            "description": "Human-readable description of the event"
          },
          "event_type": {
            "type": "string",
            "description": "The type of event (e.g., \"OutputDetected\", \"TransactionConfirmed\")"
          },
          "id": {
            "type": "integer",
            "format": "int64",
            "description": "Unique identifier for this event"
          }
        }
      },
...
    },
    "securitySchemes": {
      "api_key": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "Wallet API token, sent as 'X-API-Key: <token>'"
      },
      "bearer_token": {
        "type": "http",
        "scheme": "bearer",
        "description": "Wallet API token, sent as 'Authorization: Bearer <token>'"
      }
    }
  },
  "security": [
    {
      "bearer_token": []
    },
    {
      "api_key": []
    }
  ],
...
}